Make script to extract 16S rRNA subregions, update Makefile, close #11.

+++ code/get_rrnDB_files.sh
@@ -11,5 +11,13 @@ path=`echo $target | sed -E "s_(.*/).*_\1_"`
 
 curl -kL https://rrndb.umms.med.umich.edu/static/download/"$filename".zip -o "$target".zip
 unzip -n -d "$path" "$target".zip
-touch "$target"
-rm "$target".zip
+
+# if the last operation was successful, update timestamp
+if [[ $? -eq 0 ]]
+then
+  touch "$target"
+  rm "$target".zip
+else
+  echo "ERROR: Unable to successfully extract $filename."
+  exit 1
+fi
